feat: add a body field to procedure
The sole difference between a declaration of a procedure type and a
definition of a procedure is an existence of a body (which may be
either a block or the uninitialized token). To make it easier to
decide which of the two we are dealing with, add a field for the
procedure body, so we can query its existence.

There is also a way to differentiate them through the "#type"
directive, but it is optional, so it is not used everywhere.